refactor(step2): lotto controller로부터 lotto view 의존성 분리
